+++ src/main/java/Tasklist.java
@@ -1,0 +1,12 @@
+import java.util.ArrayList;
+public class Tasklist {
+
+    public Tasklist()
+    private static ArrayList<Task> tasks = new ArrayList<>();
+    private static int taskCount = 0;
+
+    private void addTaskCount() {
+        taskCount++;
+    }
+
+}
